+++ HISTORY.rst
@@ -4,6 +4,7 @@ History
 =======
 v0.7.7
 ------
+* Feature: Add track to user playlist, user tracks from ISRC (#96) - tehkillerbee_
 * Feature: Add support for moving playlist items (#116) - tehkillerbee_
 * Feature: Allow adding items multiple times to the same playlist - tehkillerbee_
 * Feature: Add support for adding items to a playlists at a specific position (#116) - tehkillerbee_